Fixed main.py issues and made changes
Found and fixed logic errors, more error checking using sanInput, etc. in main.py
i have also changed the stat system partially and removed the saves location file

+++ playerFile.py
@@ -5,190 +5,163 @@ from loggingConfig import initLogger
 from ast import literal_eval
 initLogger(filePath)
 class PlayerClass: # has all of the stats for the player
-    def __init__(self):
-        self.name = ""
-        self.race = {}
-        self.health = 20
-        self.level = 0
-        self.xp = 0
-        self.sparePoints = 0
-        self.gold = 0
-        self.dexterity = 1
-        self.agility = 1
-        self.vitality = 1
-        self.awareness = 1
-        self.charisma = 1
-        self.intelligence = 1
-        self.strength = 1
-        self.traits = [""]
-        self.profs = [""]
-        self.storyLocation = [""]
-        self.inventory = [""]
+    def __init__(self, 
+                 name="", 
+                 race={},
+                 health=20,
+                 level=0,
+                 xp=0,
+                 sparePoints=0,
+                 gold=0,
+                 stats={"dexterity":1,
+                        "agility":1,
+                        "vitality":1,
+                        "awareness":1,
+                        "charisma":1,
+                        "intelligence":1,
+                        "strength":1},
+                 traits = [],
+                 profs = [],
+                 storyLocation = [],
+                 inventory = []
+                 ):
+
+        self.name = name
+        self.race = race
+        self.health = health
+        self.level = level
+        self.xp = xp
+        self.sparePoints = sparePoints
+        self.gold = gold
+        self.stats = stats
+        self.traits = traits
+        self.profs = profs
+        self.storyLocation = storyLocation
+        self.inventory = inventory
     def __str__(self): # allows printing of a data sheet
+        bs = "\n\t\t"
         return f"""
-        Name: {self.name}
-        Health: {self.health}
-        Race: {self.race['name'].capitalize()}
-        Level: {self.level}
-        XP: {self.xp}
-        Gold: {self.gold}
-        SP: {self.sparePoints}
-        Dexterity: {self.dexterity}
-        Agility: {self.agility}
-        Vitality: {self.vitality}
-        Awareness: {self.awareness}
-        Charisma: {self.charisma}
-        Intelligence: {self.intelligence}
-        Strength: {self.strength}"""
+\t\tName: {self.name}
+\t\tHealth: {self.health}
+\t\tRace: {self.race['name'].capitalize()}
+\t\tLevel: {self.level}
+\t\tXP: {self.xp}
+\t\tGold: {self.gold}
+\t\tSP: {self.sparePoints}
+\t\t{bs.join((statName + ": " + str(self.stats[statName])) for statName in self.stats)}"""
     def statAssign(self):
-        stats = ["dexterity","agility","vitality","awareness","charisma","intelligence","strength"] # list of stats
-        try:
-            navigate = int(input(f"""
-            1.Dexterity - {self.dexterity}
-            2.Agility - {self.agility}
-            3.Vitality - {self.vitality}
-            4.Awareness - {self.awareness}
-            5.Charisma - {self.charisma}
-            6.Intelligence - {self.intelligence}
-            7.Strength - {self.strength}
-            8.Exit
-            You have {self.sparePoints} to spend.
-            """)) - 1
-            if navigate == 7:
-                clear("i")
-                print(self)
-                input("Input any key to continue\n>")
-                return
-            elif navigate < 0:
-                raise ValueError
-            statUp = stats[navigate]        
-        except(ValueError,IndexError): # catches unexpected values
-            print("Enter a number between 1-6")
-            clear("d")
-            self.statAssign()
+        bs = "\n\t\t"
+        enumeratedStatNames = enumerate(list(self.stats.keys()))
+        StatValues = list(self.stats.values())
+        navigate = sanInput(f"""\t\t--Choose a stat--
+\t\t{bs.join((f"{ind+1}. "+statName.capitalize()+ " - " + str(self.stats[statName])) for ind,statName in enumeratedStatNames)}
+\t\t8.Exit
+\t\tYou have {self.sparePoints} to spend.
+\t\t> """, int, 1, 8, Clear=True) - 1
+        if navigate == 7:
+            clear("i")
+            print(self)
+            input("\t\tPress enter to continue")
             return
+        statUp = list(self.stats.keys())[navigate]        
         clear("i")
-        while True:
-            try:
-                increment = int(input(f"""You have {self.sparePoints} points to spend.
-        How much would you like to increase {statUp} by?\n>""")) # gets the variable for the stat that wants increasing
-            except(ValueError):
-                print("Enter an integer between 1-7")
-                continue
-            if increment > self.sparePoints:
-                print(f"Not enough stat points! You have {self.sparePoints} points not {increment}")    
-                self.statAssign()
-                return
-            else:
-                newValue = increment + self.__dict__[statUp]  
-                break
-        while True:
-            confirm = input(f"Confirm you want to increase {statUp} from {self.__dict__[statUp]} to {newValue} (y/n)\n>").lower()
-            if confirm == "y":
-                self.__dict__[statUp] = newValue
-                self.sparePoints -= increment
-                print(f"Increased {statUp} to {newValue}")
-                clear("d")
-                return
-            elif confirm == "n":
-                print("Changes aborted")
-                clear("d")
-                self.statAssign()
-            else:
-                print("Invalid Input, please enter Y or N")
-                clear("d")
-                return
+        increment = sanInput(f"""
+\t\tYou have {self.sparePoints} points to spend.
+\t\tHow much would you like to increase {statUp.capitalize()} by?\n\t\t> """, int, 0, self.sparePoints, Clear=True) # gets the variable for the stat that wants increasing
+        newValue = increment + self.stats[statUp]  
+        confirm = sanInput(f"\t\tConfirm you want to increase {statUp} from {self.stats[statUp]} to {newValue} (y/n)\n\t\t> ", str, values=["y","n"],Clear=True)
+        if confirm == "y":
+            self.stats[statUp] = newValue
+            self.sparePoints -= increment
+            print(f"\t\tIncreased {statUp} to {newValue}")
+            clear("d")
+        elif confirm == "n":
+            print("\t\tChanges aborted")
+            clear("d")
+        return self.statAssign()
     def assignRace(self,davacis,racesDict):
         races = ["human","elf","lizardman","dwarf"]
-        raceRaw = input("""
-        What race are you?
-        1.Human
-        2.Elf
-        3.Lizardman
-        4.Dwarf
-        >""")
+        raceRaw = sanInput("""
+\t\tWhat race are you?
+\t\t1.Human
+\t\t2.Elf
+\t\t3.Lizardman
+\t\t4.Dwarf
+\t\t> """, int, 1, 4, Clear=True)
         match raceRaw:
-            case "1":
+            case 1:
                 print("""
---------Human--------
------Stat Changes----
-Intelligence + 2
-Starting Gold + 10
--Intrinsic Abilities-
-Adaptive Mind - 15% more XP gain
-----Proficiencies----
-Adaptable
-Tactics
-Ambitous""")
-            case "2":
+\t\t--------Human--------
+\t\t-----Stat Changes----
+\t\tIntelligence + 2
+\t\tStarting Gold + 10
+\t\t-Intrinsic Abilities-
+\t\tAdaptive Mind - 15% more XP gain
+\t\t----Proficiencies----
+\t\tAdaptable
+\t\tTactics
+\t\tAmbitous""")
+            case 2:
                 print("""
----------Elf---------
------Stat Changes----
-Agility + 2
-Dexterity + 2
-Awareness + 1
-Strength - 2
--Intrinsic Abilities-
-Sharp Vision - +2 to perception checks
-Darkvision - Unimpeded by darkness
-----Proficiencies---
-Archery
-Light Armour
-Foraging""") 
-            case "3":
+\t\t---------Elf---------
+\t\t-----Stat Changes----
+\t\tAgility + 2
+\t\tDexterity + 2
+\t\tAwareness + 1
+\t\tStrength - 2
+\t\t-Intrinsic Abilities-
+\t\tSharp Vision - +2 to perception checks
+\t\tDarkvision - Unimpeded by darkness
+\t\t----Proficiencies---
+\t\tArchery
+\t\tLight Armour
+\t\tForaging""") 
+            case 3:
                 print("""
-------Lizardman------
------Stat Changes----
-Base Health + 3
-Dexterity + 1
-Intelligence - 1
-Disadvantage in Dry Biomes
--Intrinsic Abilities-
-Scales - +2 defence against 'sharp' attacks
------Proficiencies----
-Spears""")            
-            case "4":
+\t\t------Lizardman------
+\t\t-----Stat Changes----
+\t\tBase Health + 3
+\t\tDexterity + 1
+\t\tIntelligence - 1
+\t\tDisadvantage in Dry Biomes
+\t\t-Intrinsic Abilities-
+\t\tScales - +2 defence against 'sharp' attacks
+\t\t-----Proficiencies----
+\t\tSpears""")            
+            case 4:
                 print("""
---------Dwarf---------
------Stat Changes-----
-Strength + 2
-Awareness + 1
-Speed - 1
--Intrinsic Abilities--
-Darkvision - Unimpeded by darkness
------Proficiencies----
-Heavy Armour
-Blunt""")
-            case other:
-                print("Unable to find race, please enter an integer between 1-4")
-                clear("d")
-                self.assignRace(davacis,racesDict)
-                return 
-        raceRaw = int(raceRaw) - 1
-        self.race = racesDict[f"{races[int(raceRaw)]}"]
+\t\t--------Dwarf---------
+\t\t-----Stat Changes-----
+\t\tStrength + 2
+\t\tAwareness + 1
+\t\tSpeed - 1
+\t\t-Intrinsic Abilities--
+\t\tDarkvision - Unimpeded by darkness
+\t\t-----Proficiencies----
+\t\tHeavy Armour
+\t\tBlunt""")
+        raceRaw = raceRaw - 1
+        self.race = racesDict[f"{races[raceRaw]}"]
         self.raceDavacisLoad()
     def raceDavacisLoad(self):
-        confirm = input("Are you sure you want to choose this race? (y/n)\n>").lower()
+        confirm = sanInput("\t\tAre you sure you want to choose this race? (y/n)\n\t\t>", str, values=["y", "n"])
         if confirm == "y":
             davacisList = self.race["davacis"]
             for i in range(len(davacisList)):
-                self.__dict__[davacis[i]] += davacisList[i]
+                self.stats[davacis[i]] += davacisList[i]
         elif confirm == "n":
             clear("i")
             self.assignRace(davacis,racesDict)
             return
-        else:
-            print("Invalid input")
-            clear("d")
     def makeSaveSlot(self,savePath):
-        saveSlot = input("What would you like to name the save slot? >").strip().replace(' ', '_')
+        saveSlot = input("\t\tWhat would you like to name the save slot?\n\t\t> ").strip().replace(' ', '_')
         if "\\" in saveSlot or "/" in saveSlot:
-            print("Save slot cannot contain slashes")
+            print("\t\tSave slot cannot contain slashes")
             clear("d")
             self.makeSaveSlot(savePath)
             return
         elif saveSlot == "":
-            print("Slot name cannot be empty")
+            print("\t\tSlot name cannot be empty")
             clear("d")
             self.makeSaveSlot(savePath)
             return
@@ -196,21 +169,19 @@ Blunt""")
         try:
             os.mkdir(f"{slotPath}")
         except(FileExistsError):
-            resolved = False
-            while not resolved:
-                if len(os.listdir(f"{slotPath}")) == 0:
-                    logging.warning("Slot creation failed due to empty slot taking name.")
-                    navigate = input("Slot name is taken by an (appearingly) empty slot. Override? (y/n)\n>").lower()
-                    if navigate == "y":
-                        os.rmdir(slotPath)
-                        resolved = True
-                    elif navigate == "n":
-                        print("Returning to start of slot naming.")
-                        clear("d")
-                    else:
-                        print("Invalid Input")
+
+            if len(os.listdir(f"{slotPath}")) == 0:
+                logging.warning("Slot creation failed due to empty slot taking name.")
+                navigate = sanInput("Slot name is taken by an (appearingly) empty slot. Override? (y/n)\n> ",str,values=["y","n"], Clear=True)
+                if navigate == "y":
+                    os.rmdir(slotPath)
+                elif navigate == "n":
+                    print("Returning to start of slot naming.")
+                    return self.makeSaveSlot(savePath)
+                    
+
             logging.warning("Slot name taken.")
-            override = input("Would you like to override the slot? (y/n)\n>").lower()
+            override = sanInput("Would you like to override the slot? (y/n)\n> ", str, values=["y","n"], Clear=True)
             if override == "y":
                 with open(f"{savePath}\\{saveSlot}\\player.txt","w"):pass
             elif override == "n":
@@ -229,61 +200,53 @@ Blunt""")
         self.saveGame(slotPath)
     def saveGame(self,slotPath):
         slotContent = os.listdir(f"{slotPath}")
+        quote = '\''
         if "player.txt" in slotContent:
             print("Saving...")
             for attribute, value in self.__dict__.items():    
                 with open(f"{slotPath}/player.txt","a") as f:
-                    f.write(f"{attribute}={value}\n")
+                    f.write(f"{attribute}={(value if (not type(value) == str) else (quote + value + quote))}\n")
         else:
             with open(f"{slotPath}/player.txt","x") as f:
                 self.saveGame(slotPath)
         clear("d")    
     def loadGame(self,savePath):
-            print("What slot would you like to load")
             slots = os.listdir(savePath)
+            message = "What slot would you like to load?\n"
             for i in range(len(slots)):
-                print(f"{i + 1}.{slots[i]}")
-            print(f"{len(slots) + 1}.Exit")
-            try:
-                saveSlot = slots[1 - int(input(">"))]
-                slotPath = f"{savePath}/{saveSlot}"
-                if len(os.listdir(f"{slotPath}")) == 0:
-                    logging.warning(f"{saveSlot} appears empty, this indicates a broken save.")
-                    delete = input("Would you like to delete the file? (y/n)\n>").lower()
-                    if delete == "y":
-                        os.rmdir(slotPath)
-                        logging.debug(f"{saveSlot} deleted")
-                    else:
-                        clear("d")
-                        self.loadGame(savePath)
-                        return
-            except(ValueError,IndexError):
-                logging.debug("User entered non-integer")
-                print("Please enter an integer listed on the left.")
-                clear("d")   
-                self.loadGame(savePath)
-                return
+                message+=f"{i + 1}.{slots[i]}\n"
+            message+=f"{len(slots) + 1}.Exit\n> "
+            choice = sanInput(message, int, 1, len(slots) + 1, Clear=True)
+            if choice == len(slots) + 1:
+                return "CONTINUE"
+
+            saveSlot = slots[choice-1]
+            slotPath = f"{savePath}/{saveSlot}"
+            if len(os.listdir(f"{slotPath}")) == 0:
+                logging.warning(f"{saveSlot} appears empty, this indicates a broken save.")
+                delete = sanInput("Would you like to delete the file? (y/n)\n> ", str, values=["y", "n"],Clear=True)
+                if delete == "y":
+                    os.rmdir(slotPath)
+                    logging.debug(f"{saveSlot} deleted")
+                else:
+                    clear("d")
+                    self.loadGame(savePath)
+                    return
+
             with open(f"{slotPath}/player.txt") as f:    
-                lines = f.readlines()
-                playerLoading = {}
+                lines = f.read().split("\n")
                 for lineNum in range(len(lines)):
                     content = lines[lineNum].split("=")
-                    key = content[0];value = content[1]
-                    playerLoading[key] = value
-                for key in playerLoading:
-                    setattr(self, key, playerLoading[key])
-                self.race = dict(literal_eval(f"{self.race}"))
-                while True:
-                    print(self)
-                    navigate = str(input(f"{self}\nIs this the correct file?(y/n)\n>")).lower()
-                    if navigate == "n":
-                        print("Returning to slot selection...")
-                        clear("d")
-                        self.loadGame(savePath)
-                        return
-                    elif navigate == "y":
-                        return
-                    else:
-                        print("Invalid input, input either 'Y' or 'N'")
+                    key = content[0];value = content[1] if len(content) >= 2 else None
+                    if key.replace(" ","") != "":
+                        setattr(self, key, literal_eval(value))
+                bs = "\n"
+                navigate = sanInput(f"{self}\n\t\tIs this the correct file?(y/n)\n\t\t> ",str, values=["y", "n"], Clear=True)
+                if navigate == "n":
+                    print("Returning to slot selection...")
+                    clear("d")
+                    self.loadGame(savePath)
+
+                return
     def nameSelf(self):
         self.name = str(input("What is your name?\n>"))
